+++ src/print_tasks/D110PrintTask.ts
@@ -15,7 +15,7 @@ export class D110PrintTask extends AbstractPrintTask {
   }
 
   override printPage(image: EncodedImage, quantity?: number): Promise<void> {
-    this.checkAddPage(quantity ?? 1);
+    this.validatePage(image, quantity ?? 1);
 
     return this.abstraction.sendAll(
       [